requirements.txt
apache-airflow==2.6.0
pandas
psycopg2-binary
requests
